+++ vnpy/trader/ui/mainwindow.py
@@ -41,7 +41,7 @@ class MainWindow(QtWidgets.QMainWindow):
         self.main_engine: MainEngine = main_engine
         self.event_engine: EventEngine = event_engine
 
-        self.window_title: str = f"Veighna Trader 社区版 - {vnpy.__version__}   [{TRADER_DIR}]"
+        self.window_title: str = f"VeighNa Trader 社区版 - {vnpy.__version__}   [{TRADER_DIR}]"
 
         self.widgets: Dict[str, QtWidgets.QWidget] = {}
 
@@ -303,7 +303,7 @@ class MainWindow(QtWidgets.QMainWindow):
         """
         Sending a test email.
         """
-        self.main_engine.send_email("Veighna Trader", "testing")
+        self.main_engine.send_email("VeighNa Trader", "testing")
 
     def open_forum(self) -> None:
         """